requirements.txt
langchain
chromadb
gpt4all
discord.py